Ajout fonction fusion des tableaux
